Add Site2Site Test case
