+++ libminifi/test/unit/SerializationTests.cpp
@@ -30,54 +30,6 @@
 #define FMT_DEFAULT fmt_lower
 
 using namespace org::apache::nifi::minifi::io;
-TEST_CASE("TestSetPortId", "[S2S1]") {
-
-  std::unique_ptr<minifi::Site2SitePeer> peer =
-      std::unique_ptr < minifi::Site2SitePeer
-          > (new minifi::Site2SitePeer(
-              std::unique_ptr < DataStream > (new DataStream()), "fake_host",
-              65433));
-
-  minifi::Site2SiteClientProtocol protocol(std::move(peer));
-
-  std::string uuid_str = "c56a4180-65aa-42ec-a945-5fd21dec0538";
-
-  uuid_t fakeUUID;
-
-  uuid_parse(uuid_str.c_str(), fakeUUID);
-
-  protocol.setPortId(fakeUUID);
-
-  REQUIRE(uuid_str == protocol.getPortId());
-
-}
-
-TEST_CASE("TestSetPortIdUppercase", "[S2S2]") {
-
-  std::unique_ptr<minifi::Site2SitePeer> peer =
-      std::unique_ptr < minifi::Site2SitePeer
-          > (new minifi::Site2SitePeer(
-              std::unique_ptr < DataStream > (new DataStream()), "fake_host",
-              65433));
-
-  minifi::Site2SiteClientProtocol protocol(std::move(peer));
-
-  std::string uuid_str = "C56A4180-65AA-42EC-A945-5FD21DEC0538";
-
-  uuid_t fakeUUID;
-
-  uuid_parse(uuid_str.c_str(), fakeUUID);
-
-  protocol.setPortId(fakeUUID);
-
-  REQUIRE(uuid_str != protocol.getPortId());
-
-  std::transform(uuid_str.begin(), uuid_str.end(), uuid_str.begin(), ::tolower);
-
-  REQUIRE(uuid_str == protocol.getPortId());
-
-}
-
 TEST_CASE("TestWriteUTF", "[MINIFI193]") {
 
   DataStream baseStream;
@@ -128,125 +80,3 @@ TEST_CASE("TestWriteUTF3", "[MINIFI193]") {
 
 }
 
-void sunny_path_bootstrap(SiteToSiteResponder *collector) {
-
-  char a = 0x14; // RESOURCE_OK
-  std::string resp_code;
-  resp_code.insert(resp_code.begin(), a);
-  collector->push_response(resp_code);
-
-  // Handshake respond code
-  resp_code = "R";
-  collector->push_response(resp_code);
-  resp_code = "C";
-  collector->push_response(resp_code);
-  char b = 0x1;
-  resp_code = b;
-  collector->push_response(resp_code);
-
-  // Codec Negotiation
-  resp_code = a;
-  collector->push_response(resp_code);
-}
-
-TEST_CASE("TestSiteToSiteVerifySend", "[S2S2]") {
-
-  SiteToSiteResponder *collector = new SiteToSiteResponder();
-
-  sunny_path_bootstrap(collector);
-
-  std::unique_ptr<minifi::Site2SitePeer> peer = std::unique_ptr
-      < minifi::Site2SitePeer
-      > (new minifi::Site2SitePeer(
-          std::unique_ptr < minifi::io::DataStream > (collector), "fake_host",
-          65433));
-
-  minifi::Site2SiteClientProtocol protocol(std::move(peer));
-
-  std::string uuid_str = "C56A4180-65AA-42EC-A945-5FD21DEC0538";
-
-  uuid_t fakeUUID;
-
-  uuid_parse(uuid_str.c_str(), fakeUUID);
-
-  protocol.setPortId(fakeUUID);
-
-  REQUIRE(true == protocol.bootstrap());
-
-  REQUIRE(collector->get_next_client_response() == "NiFi");
-  collector->get_next_client_response();
-  REQUIRE(collector->get_next_client_response() == "SocketFlowFileProtocol");
-  collector->get_next_client_response();
-  collector->get_next_client_response();
-  collector->get_next_client_response();
-  collector->get_next_client_response();
-  REQUIRE(collector->get_next_client_response() == "nifi://fake_host:65433");
-  collector->get_next_client_response();
-  collector->get_next_client_response();
-  REQUIRE(collector->get_next_client_response() == "GZIP");
-  collector->get_next_client_response();
-  REQUIRE(collector->get_next_client_response() == "false");
-  collector->get_next_client_response();
-  REQUIRE(collector->get_next_client_response() == "PORT_IDENTIFIER");
-  collector->get_next_client_response();
-  REQUIRE(
-      collector->get_next_client_response()
-          == "c56a4180-65aa-42ec-a945-5fd21dec0538");
-  collector->get_next_client_response();
-  REQUIRE(collector->get_next_client_response() == "REQUEST_EXPIRATION_MILLIS");
-  collector->get_next_client_response();
-  REQUIRE(collector->get_next_client_response() == "30000");
-  collector->get_next_client_response();
-  REQUIRE(collector->get_next_client_response() == "NEGOTIATE_FLOWFILE_CODEC");
-  collector->get_next_client_response();
-  REQUIRE(collector->get_next_client_response() == "StandardFlowFileCodec");
-  collector->get_next_client_response(); // codec version
-
-  // start to send the stuff
-  // Create the transaction
-  std::string transactionID;
-  std::string payload = "Test MiNiFi payload";
-  minifi::Transaction *transaction;
-  transaction = protocol.createTransaction(transactionID, minifi::SEND);
-  collector->get_next_client_response();
-  REQUIRE(collector->get_next_client_response() == "SEND_FLOWFILES");
-  std::map < std::string, std::string > attributes;
-  minifi::DataPacket packet(&protocol, transaction, attributes, payload);
-  REQUIRE(protocol.send(transactionID, &packet, nullptr, nullptr) == true);
-  collector->get_next_client_response();
-  collector->get_next_client_response();
-  std::string rx_payload = collector->get_next_client_response();
-  ;
-  REQUIRE(payload == rx_payload);
-
-}
-
-TEST_CASE("TestSiteToSiteVerifyNegotiationFail", "[S2S2]") {
-
-  SiteToSiteResponder *collector = new SiteToSiteResponder();
-
-  char a = 0xFF;
-  std::string resp_code;
-  resp_code.insert(resp_code.begin(), a);
-  collector->push_response(resp_code);
-  collector->push_response(resp_code);
-
-  std::unique_ptr<minifi::Site2SitePeer> peer = std::unique_ptr
-      < minifi::Site2SitePeer
-      > (new minifi::Site2SitePeer(
-          std::unique_ptr < minifi::io::DataStream > (collector), "fake_host",
-          65433));
-
-  minifi::Site2SiteClientProtocol protocol(std::move(peer));
-
-  std::string uuid_str = "C56A4180-65AA-42EC-A945-5FD21DEC0538";
-
-  uuid_t fakeUUID;
-
-  uuid_parse(uuid_str.c_str(), fakeUUID);
-
-  protocol.setPortId(fakeUUID);
-
-  REQUIRE(false == protocol.bootstrap());
-
-}
